Bug fixing

+++ Foksal/Forms/Agreements/FrmAgreement.cs
@@ -613,7 +613,7 @@ namespace Foksal.Forms.Agreements
             else
             {
                 MessageBox.Show("Nie masz uprawnień do wykonania tej akcji.", "Niewystarczające uprawnienia", MessageBoxButtons.OK, MessageBoxIcon.Stop);
-            }            
+            }
         }
 
         private void btnAddPosition_Click(object sender, EventArgs e)
@@ -623,16 +623,23 @@ namespace Foksal.Forms.Agreements
 
         private void btnRemovePosition_Click(object sender, EventArgs e)
         {
-            if (gridExPositions.Row > -1)
+            if (AppUser.Instance.CanWriteAgreements)
             {
-                if (MessageBox.Show("Czy na pewno chcesz usunąć zaznaczoną pozycję?\r\nTa operacja jest nieodwracalna!", "Uwaga", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning) == DialogResult.OK)
+                if (gridExPositions.Row > -1)
                 {
-                    this.SetPositionChangesPending(false);
-                    this.isPositionAddPending = false;
-                    this.gridPositionsRepo.DeleteRow(gridExPositions.CurrentRow.RowIndex);
-                    gridExPositions.Refetch();
-                    this.LoadData();
+                    if (MessageBox.Show("Czy na pewno chcesz usunąć zaznaczoną pozycję?\r\nTa operacja jest nieodwracalna!", "Uwaga", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning) == DialogResult.OK)
+                    {
+                        this.SetPositionChangesPending(false);
+                        this.isPositionAddPending = false;
+                        this.gridPositionsRepo.DeleteRow(gridExPositions.CurrentRow.RowIndex);
+                        gridExPositions.Refetch();
+                        this.LoadData();
+                    }
                 }
+            }
+            else
+            {
+                MessageBox.Show("Nie masz uprawnień do wykonania tej akcji.", "Niewystarczające uprawnienia", MessageBoxButtons.OK, MessageBoxIcon.Stop);
             }
         }
